refactor: 拆分大型设置与实验页面 (#35)
将实验表单与结果组件、设置页稳定 section 从页面协调层中提取，保持现有 ViewModel 状态、controller 所有权与回调语义不变。
